vuong

+++ app/src/main/java/com/trainh/assignmentprm/entities/Product.java
@@ -4,11 +4,11 @@ public class Product {
     private int id;
     private int image;
     private String name;
-    private double price;
+    private float price;
     private String type;
     private String description;
 
-    public Product(int image, String name, double price, String type, String description) {
+    public Product(int image, String name, float price, String type, String description) {
         this.image = image;
         this.name = name;
         this.price = price;
@@ -16,7 +16,7 @@ public class Product {
         this.description = description;
     }
 
-    public Product(int id, int image, String name, double price, String type, String description) {
+    public Product(int id, int image, String name, float price, String type, String description) {
         this.id = id;
         this.image = image;
         this.name = name;
@@ -49,11 +49,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
